ref: cleanup code 1.3
