+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -1,26 +1,25 @@
-package ru.practicum.shareit.booking.item.model;
+package ru.practicum.shareit.item.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import ru.practicum.shareit.request.ItemRequest;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
-public class Item {
+@AllArgsConstructor
+public class ItemDto {
     private long id;
     private long owner;
     private String name;
     private String description;
     private Boolean available;
-    private ItemRequest request;
+    private Long request;
 
-    public Item(long id, long owner, String name, String description, Boolean available) {
+    public ItemDto(long id, String name, String description, Boolean available, long request) {
         this.id = id;
-        this.owner = owner;
         this.name = name;
         this.description = description;
         this.available = available;
+        this.request = request;
     }
 }
