+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,25 +1,26 @@
-package ru.practicum.shareit.booking.item.dto;
+package ru.practicum.shareit.item.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import ru.practicum.shareit.request.ItemRequest;
 
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
-public class ItemDto {
+@NoArgsConstructor
+public class Item {
     private long id;
     private long owner;
     private String name;
     private String description;
     private Boolean available;
-    private Long request;
+    private ItemRequest request;
 
-    public ItemDto(long id, String name, String description, Boolean available, long request) {
+    public Item(long id, long owner, String name, String description, Boolean available) {
         this.id = id;
+        this.owner = owner;
         this.name = name;
         this.description = description;
         this.available = available;
-        this.request = request;
     }
 }
