+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,6 +1,6 @@
-package ru.practicum.shareit.booking.item.service;
+package ru.practicum.shareit.item.service;
 
-import ru.practicum.shareit.booking.item.dto.ItemDto;
+import ru.practicum.shareit.item.dto.ItemDto;
 
 import java.util.List;
